+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Refactor `Exemplar` and `AttachmentValue` to be under `metrics.data`. Note that this is a breaking change
 if you're using the `Exemplar` classes or APIs in the previous releases.
 - Add `TagMetadata` that defines the properties associated with a `Tag`.
+- Add `QueueMetricProducer` that supports pushing and buffering `Metric`s.
 
 # 0.19.0 - 2019-01-28
 - Add an artifact `opencensus-contrib-http-jetty-client` for instrumenting jetty http client. Add extractor for Jetty Client.